Fix neo4j_ustring_value() docs (#10)
The API docs for neo4j_ustring_value() suggested that the return value would always be a valid pointer, but for empty strings, the function actually returns a null pointer.

Checking the string length to guard against this must be done with neo4j_string_length(). The API docs had a typo in the name of that function.