Allow neo4j_ntostring() buffer to be NULL
Passing a null pointer as buffer to strftime() is undefined behaviour in ISO C, even though many standard library implementations allow it, treating it like a zero-length buffer.

Our documentation of neo4j_ntostring() never mentions that you can pass a null pointer as buffer. Instead, it says that a pointer must be "a buffer", implying that NULL would be a usage error. However, it looks like NULL was in fact always supported, all the way back to the beginnings of libneo4j-client.

This change special-cases zero-length buffers in the temporal *_str functions and updates the documentation to indicate that NULL is a supported value for the buffer pointer.